requirements.txt
requests
pandas
pytesseract
Pillow
python-dotenv
openai
